chore: Move area Climb object to the area package
This does not represent the entire Climb object, but a small subset of the fields necessary for displaying climbs in the Area screen.
