--- data/src/main/kotlin/com/routesearch/data/climb/Climb.kt
@@ -1,8 +1,0 @@
-package com.routesearch.data.climb
-
-data class Climb(
-  val id: String,
-  val grades: Grades?,
-  val name: String,
-  val type: Type,
-)
