+++ data/src/main/kotlin/com/routesearch/data/climb/ClimbMapper.kt
@@ -1,21 +1,10 @@
 package com.routesearch.data.climb
 
+import com.routesearch.data.area.Area
 import com.routesearch.data.local.climb.Climb as LocalClimb
 import com.routesearch.data.local.climb.Grades as LocalGrades
-import com.routesearch.data.remote.AreaQuery.Climb as RemoteClimb
 import com.routesearch.data.remote.fragment.GradesFragment as RemoteGrade
 import com.routesearch.data.remote.fragment.TypeFragment as RemoteType
-
-@JvmName("remoteClimbToClimb")
-internal fun List<RemoteClimb?>.toClimbs() = filterNotNull()
-  .map { it.toClimb() }
-
-internal fun RemoteClimb.toClimb() = Climb(
-  id = uuid,
-  grades = grades?.gradesFragment?.toGrade(),
-  name = name,
-  type = type.typeFragment.toType(),
-)
 
 internal fun RemoteGrade.toGrade() = Grades(
   yds = yds,
@@ -36,14 +25,6 @@ internal fun RemoteType.toType() = when {
   else -> throw IllegalArgumentException("Invalid type $this")
 }
 
-internal fun Climb.toLocalClimb(areaId: String) = LocalClimb(
-  id = id,
-  areaId = areaId,
-  name = name,
-  type = type.name,
-  grades = grades?.toLocalGrades(),
-)
-
 internal fun Grades.toLocalGrades() = LocalGrades(
   yds = yds,
   vScale = vScale,
@@ -53,7 +34,7 @@ internal fun Grades.toLocalGrades() = LocalGrades(
 internal fun List<LocalClimb>.toClimbs() = filterNotNull()
   .map { it.toClimb() }
 
-internal fun LocalClimb.toClimb() = Climb(
+internal fun LocalClimb.toClimb() = Area.Climb(
   id = id,
   grades = grades?.toGrades(),
   name = name,
